+++ src/Grocery.java
@@ -1,5 +1,7 @@
+import java.math.BigDecimal;
+
 public class Grocery extends Product{
     public Grocery (String name, int price, int weight) {
-        super(name, price, weight);
+        super(name, BigDecimal.valueOf(price), weight);
     }
 }
