+++ src/LargeAppliance.java
@@ -1,6 +1,8 @@
+import java.math.BigDecimal;
+
 public class LargeAppliance extends Product{
     public LargeAppliance (String name, int price, int weight)
     {
-        super(name, price, weight);
+        super(name, BigDecimal.valueOf(price), weight);
     }
 }
